requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
eagerpy==0.29.0
foolbox==3.3.1
future==0.18.2
gast==0.2.2
gitdb==4.0.6
GitPython==3.1.14
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
joblib==0.17.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.1
matplotlib==3.2.2
numpy==1.19.0
onnx==1.7.0
onnx-tf==1.5.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.0.5
Pillow==7.2.0
pip==20.2.4
pkg-resources==0.0.0
protobuf==3.13.0
PyGnuplot==0.11.16
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
scikit-learn==0.23.2
scipy==1.1.0
seaborn==0.10.1
setuptools==39.0.1
six==1.15.0
sklearn==0.0
smmap==3.0.5
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.4.0
torchattacks==2.14.5
torchvision==0.5.0
tqdm==4.61.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wheel==0.35.1
wrapt==1.12.1
zipp==3.3.1